Renamed storage to api.go and fs to storage.go
